python-server/requirements.txt
Flask==2.3.3
openai-whisper==20231117
torch
transformers
ffmpeg-python
sentencepiece 
protobuf     